# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sistema POS Ventas API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quotes:
    get:
      operationId: getQuotes
      tags: [quotes]
      summary: List all quotes/proformas
      security:
        - bearerAuth: []
      parameters:
        - name: customerId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Create a new quote/proforma (no stock reduction)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteBody"
      responses:
        "201":
          description: Quote created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    get:
      operationId: getQuoteById
      tags: [quotes]
      summary: Get quote with details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote with details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteWithDetails"
  /quotes/{id}/convert:
    post:
      operationId: convertQuoteToSale
      tags: [quotes]
      summary: Convert quote to sale (reduces stock)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Sale created from quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"

  # DASHBOARD
...
  /reports/sales-by-user:
    get:
      operationId: getReportSalesByUser
      tags: [reports]
      summary: Sales grouped by user/vendor
      security:
        - bearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          required: false
          schema:
            type: string
        - name: dateTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Sales by user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesByUserReport"
  /reports/by-category:
    get:
      operationId: getReportByCategory
      tags: [reports]
      summary: Sales/revenue grouped by category and brand
      security:
        - bearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          required: false
          schema:
            type: string
        - name: dateTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Report by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryReport"
  /reports/stock-alerts:
    get:
      operationId: getReportStockAlerts
      tags: [reports]
      summary: Products with critical or zero stock
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stock alert items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockAlertItem"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SaleItem:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
      required: [productId, quantity, unitPrice]
...
    SaleDetail:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: ["integer", "null"]
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
      required: [id, productName, quantity, unitPrice, subtotal]
    Sale:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: ["string", "null"]
        customerId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        subtotal:
          type: number
        iva:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, subtotal, iva, total, paymentMethod, status, createdAt]
...
    CreateQuoteBody:
      type: object
      properties:
        customerId:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        paymentMethod:
          type: string
        notes:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
      required: [items, paymentMethod]
    Quote:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        subtotal:
          type: number
        iva:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, subtotal, iva, total, paymentMethod, status, createdAt]
    QuoteWithDetails:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        subtotal:
          type: number
        iva:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
        createdAt:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/SaleDetail"
      required: [id, subtotal, iva, total, paymentMethod, status, createdAt, details]

    # DASHBOARD
...
    SalesByUserReport:
      type: object
      properties:
        userId:
          type: integer
        userName:
          type: string
        totalSales:
          type: integer
        totalRevenue:
          type: number
      required: [userId, userName, totalSales, totalRevenue]
    CategoryReport:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: string
        totalSold:
          type: integer
        totalRevenue:
          type: number
      required: [categoryName, totalSold, totalRevenue]
    StockAlertItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        stock:
          type: integer
        minStock:
          type: integer
        categoryName:
          type: ["string", "null"]
        brandName:
          type: ["string", "null"]
        status:
          type: string
      required: [id, name, stock, minStock, status]